Refactor of files to separate the corresponding endpoints in different controllers + UNIT TESTS + INTEGRATION TESTS
